LPS-145980 Move component's specific utils to the component folder
