+++ modules/apps/content-dashboard/content-dashboard-web/src/main/resources/META-INF/resources/js/components/SelectFileExtension.js
@@ -15,8 +15,8 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import {nodeTreeArrayMapper} from '../utils/tree-utils';
-import TreeFilter from './TreeFilter';
+import TreeFilter from './TreeFilter/TreeFilter';
+import {nodeTreeArrayMapper} from './TreeFilter/treeUtils';
 
 const SelectFileExtension = ({
 	fileExtensionGroups,
