+++ modules/apps/content-dashboard/content-dashboard-web/src/main/resources/META-INF/resources/js/components/TreeFilter/TreeFilter.js
@@ -25,7 +25,7 @@ import {
 	getSelectedNodeObject,
 	selectedDataOutputTransfomer,
 	visit,
-} from '../utils/tree-utils';
+} from './treeUtils';
 
 const SEARCH_QUERY_MIN_LENGHT = 2;
 const SEARCH_INPUT_DEBOUNCE = 300;
